+++ include/gcode_comands.cpp
@@ -1,25 +1,5 @@
 #include "gcode_comands.h"
 
-// Fills target container by X,Y,Z values
-void setTargetPosition() {
-        if (Gparser.HasWord('X'))
-          TARGET[X_AXIS] = getCommandValue('X');
-        if (Gparser.HasWord('Y'))
-          TARGET[Y_AXIS] = getCommandValue('Y');
-        if (Gparser.HasWord('Z'))
-          TARGET[Z_AXIS] = getCommandValue('Z');
-}
-
-// Checks if the command is there
-bool isCommand(char command) {
-  return Gparser.HasWord(command);
-}
-
-// Get a value after big chars
-double getCommandValue(char command){
-  //(double)Gparser.GetWordValue('S')
-  return Gparser.GetWordValue(command);
-}
 
 void G0_Move(Motor &motorX, Motor &motorY, Motor &motorZ, double *target)
 {
@@ -49,25 +29,3 @@ void G28_Move(Motor &motorX, Motor &motorY, Motor &motorZ)
   motorZ.reset();
 }
 
-int procesGCommand(int commandNumber, double *target)
-{
-  switch (commandNumber)
-  {
-
-  case 0:
-    G0_Move(target);
-    Serial.println("G0:OK");
-    break;
-  case 1:
-    G1_Move(target);
-    Serial.println("G1:OK");
-    break;
-  case 28:
-    G28_Move();
-    Serial.println("G28:OK");
-    break;
-  default:
-    break;
-  }
-  return 0;
-}
