+++ src/main.cpp
@@ -1,10 +1,9 @@
 #include <Arduino.h>
 #include <GCodeParser.h>
-#include "GyverStepper.h"
 #include <GyverPID.h>
 #include <GyverNTC.h>
 #include <GyverTimers.h>
-
+#include "GyverStepper.h"
 #include "motor.h"
 #include "tones_types.h"
 #include "config.h"
@@ -85,4 +84,5 @@ void loop()
 {
   parsing();
 }
+
